Implement fortius support based on github FortAnt project technology.
Device brake force for sim mode is now set directly via new
setSimState api on deviceController, which allows for simulations
force to be set directly on trainer, instead of each trainer
implementing their own gradient simulation.

Devices that dont implement setSimSate are called with setGradient
as before.

Remove special case for non-workout modes, they are now handled
via loadUpdate and timer, same as workout modes. This is necessary
to send force updates when no workout is present.

This change does not implement:
- setSimState for other trainers that could benefit from it.
- calibration for fortius
- virtual and device speed synchronization for fortius.

Huge thanks to Mattipee and MickyDuck for reviews, testing and
feedback.